requirements.txt
scikit-learn==0.20.2
Pillow==6.0.0 
Flask==1.0.2